requirements.txt
google-apps-meet==0.1.8
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
protobuf==5.28.0
aiogram
python-dotenv